requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Werkzeug==2.2.2
python-dotenv==1.0.1
urllib3==2.3.0
requests==2.32.3
jinja2==3.1.5
Mako==1.3.9
alembic==1.14.1
bs4
bleach
redis
flask_socketio
pytest==8.3.5
pillow
pytz